fix(workspace): narrow every file-declared length to a pointer the target holds
`usize` is 32 bits wide on the browser targets the fieldglass-wasm milestone
aims at. A `u64` length read out of a file and narrowed with a bare `as usize`
wraps there — into a value small enough to pass the bounds check that follows
it, so the reader takes the wrong bytes and answers with the wrong field.
Nothing in the project ran on a 32-bit target, so none of it was tested.

Clippy's pointer-width lints flag 83 lines, not the 59 the issue recorded.
Thirty-four now go through a checked conversion; the other forty-nine name the
bound that makes them exact in a comment beside them.

`fieldglass_core::bytes::checked_usize` is the one narrowing, in the module
whose `ByteRange` doc already said converting to `usize` is the reader's
business "at the point it has bounded the value". `Cursor::usize` and
`read_usize_le` are the HDF5-side spellings of it; netcdf's ad-hoc
`usize::try_from` sites and its `nonneg_to_usize` fold into it.

Three were wrong on any target and are fixed rather than annotated:

- `SpatialIndex::new` built its `Vec<u32>` tree over `n as u32`, so a grid with
  more cells than a `u32` counts indexed a silent prefix of itself.
- GRIB1 simple packing compared `bds.len() as u32` against `section_len`, so a
  body past 4 GiB read as *shorter* than its own header declared.
- The point probe narrowed a 64-bit grid index to `u32` before the off-grid
  guard, so an index of exactly 2^32 passed it — and on a 32-bit target went on
  to report column zero.

`MessageMeta.offset_bytes` also stops being an `i32`, which went negative past
2 GiB. It is a JS number either way, so no TypeScript changes.

+++ crates/fieldglass-grib1/src/packing/simple.rs
@@ -28,7 +28,10 @@ impl Grib1Packing for SimplePacking {
         expected_count: usize,
         _runs: StoredRuns<'_>,
     ) -> Result<Vec<Option<f64>>, FieldglassError> {
-        if (bds.len() as u32) < header.section_len {
+        // Widen rather than narrow: `section_len` is a 3-byte field, so this
+        // comparison is exact on any target, where `bds.len() as u32` would
+        // have wrapped a body longer than 4 GiB into "too short".
+        if bds.len() < header.section_len as usize {
             return Err(FieldglassError::Parse(format!(
                 "BDS body shorter than declared section_len {}",
                 header.section_len
